Added: - Dilithium - Dilithium Rails. - Dilithium Booster. - Dilithium Stabilizer. - Crystal Chamber - Started working on the Dimensional Anchor multiblock. - Added interfaces for boostable and multiblock machines. - Added util/crafting component items. - Ability chips can now be found inside labyrinths.

+++ src/main/java/sunsetsatellite/signalindustries/blocks/BlockCrusher.java
@@ -47,11 +47,11 @@ public class BlockCrusher extends BlockContainerTiered{
                         }
 
                         itemstack.stackSize -= i1;
-                        EntityItem entityitem = new EntityItem(world, (double) ((float) i + f), (double) ((float) j + f1), (double) ((float) k + f2), new ItemStack(itemstack.itemID, i1, itemstack.getMetadata()));
+                        EntityItem entityitem = new EntityItem(world, (float) i + f, (float) j + f1, (float) k + f2, new ItemStack(itemstack.itemID, i1, itemstack.getMetadata()));
                         float f3 = 0.05F;
-                        entityitem.motionX = (double) ((float) random.nextGaussian() * f3);
-                        entityitem.motionY = (double) ((float) random.nextGaussian() * f3 + 0.2F);
-                        entityitem.motionZ = (double) ((float) random.nextGaussian() * f3);
+                        entityitem.motionX = (float) random.nextGaussian() * f3;
+                        entityitem.motionY = (float) random.nextGaussian() * f3 + 0.2F;
+                        entityitem.motionZ = (float) random.nextGaussian() * f3;
                         world.entityJoinedWorld(entityitem);
                     }
                 }
